+++ client/src/utils/getWeatherType.js
